Criando metodos, mapper e dtos.

+++ backend/ms-publicacoes/src/main/java/br/com/elastech/ms_publicacoes/dto/response/CriarPublicacaoResponse.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record CriarPublicacaoResponse(
-        Long usuarioId,
+        Integer usuarioId,
         String conteudo,
         String imagem
 ) {
